Move modelType class to parser
Moved ModelExtractor class to parser.
