+++ src/main/java/seedu/duke/controller/ControlManager.java
@@ -4,6 +4,7 @@ import seedu.duke.Duke;
 import seedu.duke.common.Messages;
 import seedu.duke.controller.command.FindCommand;
 import seedu.duke.controller.command.ListCommand;
+import seedu.duke.controller.parser.ModelExtractor;
 import seedu.duke.controller.parser.ModelParser;
 import seedu.duke.exception.ContactParamException;
 import seedu.duke.exception.EmptyParameterException;
@@ -19,7 +20,7 @@ import seedu.duke.controller.command.Command;
 import seedu.duke.controller.command.CommandFactory;
 import seedu.duke.controller.parser.CommandParser;
 import seedu.duke.controller.parser.CommandType;
-import seedu.duke.model.ModelType;
+import seedu.duke.controller.parser.ModelType;
 import seedu.duke.model.event.Event;
 import seedu.duke.model.quiz.Quiz;
 import seedu.duke.storage.EventStorageManager;
