+++ src/main/java/seedu/duke/controller/parser/ModelType.java
@@ -1,0 +1,8 @@
+package seedu.duke.controller.parser;
+
+/**
+ * Represents all possible models available to the user.
+ */
+public enum ModelType {
+    CLASS, CCA, TEST, TUITION, QUIZ, CONTACT, EVENT;
+}
