--- src/main/java/seedu/duke/model/ModelType.java
@@ -1,5 +1,0 @@
-package seedu.duke.model;
-
-public enum ModelType {
-    CLASS, CCA, TEST, TUITION, QUIZ, CONTACT, EVENT;
-}
