+++ src/test/java/seedu/duke/controller/ModelExtractorTest.java
@@ -1,10 +1,11 @@
 package seedu.duke.controller;
 
 import org.junit.jupiter.api.Test;
+import seedu.duke.controller.parser.ModelExtractor;
 import seedu.duke.exception.InvalidModelException;
 import seedu.duke.model.DataManager;
 import seedu.duke.model.Model;
-import seedu.duke.model.ModelType;
+import seedu.duke.controller.parser.ModelType;
 import seedu.duke.model.contact.ContactManager;
 import seedu.duke.model.event.EventManager;
 import seedu.duke.model.event.EventParameter;
